requirements.txt
numpy
pillow
drawing
network